refactor: rearrange folder structure on container Rearrange folder structure on container to follow single responsibility principle. Add code comments on container workspace.
