+++ packages/container/src/pages/404.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+/**
+ *
+ * @returns 404 Page.
+ */
+export const Page404 = (): JSX.Element => (
+  <div>404 Page not found</div>
+);
